redefine the llm func call struct
